Move header file generation logic
WIP, still need to get the length of the metadat object, but otherwise
everything should be working.
