+++ lib/engine-object-file/src/artifact.rs
@@ -237,174 +237,6 @@ impl ObjectFileArtifact {
         Self::from_parts_crosscompiled(&mut *engine_inner, metadata, obj_bytes, metadata_length)
     }
 
-    /// Generate the header file that goes with the generated object file.
-    pub fn generate_header_file(&self) -> String {
-        use crate::header::*;
-        let mut c_statements = vec![];
-        c_statements.push(CStatement::Declaration {
-            name: "module_bytes_len".to_string(),
-            array: false,
-            is_extern: false,
-            is_const: true,
-            ctype: CType::U32,
-            definition: Some(Box::new(CStatement::LiteralConstant {
-                value: format!("{}", self.metadata_length),
-            })),
-        });
-        c_statements.push(CStatement::Declaration {
-            name: "WASMER_METADATA".to_string(),
-            array: true,
-            is_extern: true,
-            is_const: true,
-            ctype: CType::U8,
-            definition: None,
-        });
-        for (function_local_index, _function_len) in self.metadata.function_body_lengths.iter() {
-            let function_name = self
-                .metadata
-                .symbol_to_name(Symbol::LocalFunction(function_local_index));
-            // TODO: figure out the signature here too
-            c_statements.push(CStatement::Declaration {
-                name: function_name.clone(),
-                array: false,
-                is_extern: false,
-                is_const: false,
-                ctype: CType::Function {
-                    arguments: vec![CType::Void],
-                    return_value: None,
-                },
-                definition: None,
-            });
-        }
-
-        // function pointer array
-        {
-            let mut function_pointer_array_statements = vec![];
-            for (function_local_index, _function_len) in self.metadata.function_body_lengths.iter()
-            {
-                let function_name = self
-                    .metadata
-                    .symbol_to_name(Symbol::LocalFunction(function_local_index));
-                // TODO: figure out the signature here too
-
-                function_pointer_array_statements.push(CStatement::LiteralConstant {
-                    value: function_name.clone(),
-                });
-            }
-
-            c_statements.push(CStatement::Declaration {
-                name: "function_pointers".to_string(),
-                array: true,
-                is_extern: false,
-                is_const: true,
-                ctype: CType::void_ptr(),
-                definition: Some(Box::new(CStatement::LiteralArray {
-                    items: function_pointer_array_statements,
-                })),
-            });
-        }
-
-        for (sig_index, _func_type) in self.metadata.compile_info.module.signatures.iter() {
-            let function_name = self
-                .metadata
-                .symbol_to_name(Symbol::FunctionCallTrampoline(sig_index));
-
-            c_statements.push(CStatement::Declaration {
-                name: function_name.clone(),
-                array: false,
-                is_extern: false,
-                is_const: false,
-                ctype: CType::Function {
-                    arguments: vec![CType::void_ptr(), CType::void_ptr(), CType::void_ptr()],
-                    return_value: None,
-                },
-                definition: None,
-            });
-        }
-
-        // function trampolines
-        {
-            let mut function_trampoline_statements = vec![];
-            for (sig_index, _vm_shared_index) in self.metadata.compile_info.module.signatures.iter()
-            {
-                let function_name = self
-                    .metadata
-                    .symbol_to_name(Symbol::FunctionCallTrampoline(sig_index));
-                function_trampoline_statements.push(CStatement::LiteralConstant {
-                    value: function_name,
-                });
-            }
-
-            c_statements.push(CStatement::Declaration {
-                name: "function_trampolines".to_string(),
-                array: true,
-                is_extern: false,
-                is_const: true,
-                ctype: CType::void_ptr(),
-                definition: Some(Box::new(CStatement::LiteralArray {
-                    items: function_trampoline_statements,
-                })),
-            });
-        }
-
-        for func_index in self
-            .metadata
-            .compile_info
-            .module
-            .functions
-            .keys()
-            .take(self.metadata.compile_info.module.num_imported_functions)
-        {
-            let function_name = self
-                .metadata
-                .symbol_to_name(Symbol::DynamicFunctionTrampoline(func_index));
-            // TODO: figure out the signature here
-            c_statements.push(CStatement::Declaration {
-                name: function_name,
-                array: false,
-                is_extern: false,
-                is_const: false,
-                ctype: CType::Function {
-                    arguments: vec![CType::void_ptr(), CType::void_ptr(), CType::void_ptr()],
-                    return_value: None,
-                },
-                definition: None,
-            });
-        }
-
-        // dynamic function trampoline pointer array
-        {
-            let mut dynamic_function_trampoline_statements = vec![];
-            for func_index in self
-                .metadata
-                .compile_info
-                .module
-                .functions
-                .keys()
-                .take(self.metadata.compile_info.module.num_imported_functions)
-            {
-                let function_name = self
-                    .metadata
-                    .symbol_to_name(Symbol::DynamicFunctionTrampoline(func_index));
-                dynamic_function_trampoline_statements.push(CStatement::LiteralConstant {
-                    value: function_name,
-                });
-            }
-            c_statements.push(CStatement::Declaration {
-                name: "dynamic_function_trampoline_pointers".to_string(),
-                array: true,
-                is_extern: false,
-                is_const: true,
-                ctype: CType::void_ptr(),
-                definition: Some(Box::new(CStatement::LiteralArray {
-                    items: dynamic_function_trampoline_statements,
-                })),
-            });
-        }
-
-        generate_c(&c_statements)
-    }
-
     /// Get the default extension when serializing this artifact
     pub fn get_default_extension(triple: &Triple) -> &'static str {
         match triple.operating_system {
@@ -639,5 +471,9 @@ impl Artifact for ObjectFileArtifact {
     /// Serialize a ObjectFileArtifact
     fn serialize(&self) -> Result<Vec<u8>, SerializeError> {
         Ok(self.module_bytes.clone())
+    }
+
+    fn symbol_registry(&self) -> &dyn SymbolRegistry {
+        &self.metadata
     }
 }
