+++ lib/engine-object-file/src/lib.rs
@@ -24,7 +24,6 @@
 mod artifact;
 mod builder;
 mod engine;
-pub(crate) mod header;
 mod serialize;
 
 pub use crate::artifact::ObjectFileArtifact;
